% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcache.R
\name{qcache}
\alias{qcache}
\title{Cache the results of a function call}
\usage{
qcache(cache_dir, base_name, fn = NULL, ...)
}
\arguments{
\item{cache_dir}{Full path to directory where .Rds file will be stored.}

\item{base_name}{Basename of .Rds file (i.e. ".Rds" will be appended).}

\item{fn}{A function or name of function.}

\item{...}{Additional parameters passed to \code{fn}.}
}
\value{
The result of the function call with parameters \code{...}
}
\description{
Cache the results of a function call as an .Rds file with a specified
basename, unless there is already a cached version stored, in which case load
the cached version.
}
\details{
Remove a cached object by deleting the file manually or calling
\code{\link{unqcache}}.
}
\examples{
# cache the result of this call
qcache("~/cache", "rnorm_result", rnorm, n = 100, mean = 0, sd = 1)

# load the cached version
qcache("~/cache", "rnorm_result", rnorm, n = 100, mean = 0, sd = 1)
# alternatively,
qcache("~/cache", "rnorm_result")
}
