% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqcache.R
\name{rqcache}
\alias{rqcache}
\title{Read cached function call results}
\usage{
rqcache(cache_dir, base_name)
}
\arguments{
\item{cache_dir}{Full path to directory where .Rds file will be stored}

\item{base_name}{Vector of .Rds basenames}
}
\value{
Cached object
}
\description{
Read cached function call results, i.e. an .Rds file in the specified
\code{cache_dir} created by \code{\link{qcache}}.
}
\examples{
rqcache("~/cache", "rnorm_result")
}
