% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unqcache.R
\name{unqcache}
\alias{unqcache}
\title{Remove cached function call results}
\usage{
unqcache(cache_dir, base_names = NULL)
}
\arguments{
\item{cache_dir}{Full path to directory where .Rds file will be stored}

\item{base_names}{Vector of .Rds basenames}
}
\value{
Vector of basenames successfully deleted from file system
}
\description{
Remove cached function call results, i.e. .Rds files in the specified
\code{cache_dir} created by \code{\link{qcache}}.
}
\details{
If \code{base_names} is NULL, all cached files in \code{cache_dir} are
removed, otherwise only those matching the vector of basenames are removed.
}
\examples{
unqcache("~/cache", c("rnorm_result", "runif_result"))
}
